chore: use TaksContext in TaskDetail component and subcomponents (ViewMode & EditMode)
