+++ frontend/src/context/TaskContext.jsx
@@ -24,8 +24,9 @@ export default function TaskProvider({children}) {
   const [page, setPage] = useState(1);
   const [showDiscarded, setShowDiscarded] = useState(false);
 
-  // single task related states
+  // single task states
   const [selectedTask, setSelectedTask] = useState(null);
+  const [isEditing, setIsEditing] = useState(false);
 
 
   // List-related operations
@@ -52,16 +53,65 @@ export default function TaskProvider({children}) {
   }, [page, showDiscarded]);
 
 
-  // Singe-task operations
+  const toggleDiscardedView = () => {
+    setShowDiscarded(!showDiscarded);
+    setPage(1);
+  }
 
+
+  // Single-task operations
+
+  const updateTask = async(taskId, updatedData) => {
+    await updateDatabase(taskId, updatedData);
+    await fetchTasks();
+
+    setSelectedTask(prev => ({
+      ...prev,
+      ...updatedData // overwrite the previous task details
+    }));
+
+    setIsEditing(false);
+  }
+
+
+  // Update DataBase
+  const updateDatabase = async (id, updatedData) => {
+    const token = localStorage.getItem('token');
+
+    await fetch(getApiUrl(`/api/task/edit/${id}`), {
+      method: 'PUT',
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${token}`,
+      },
+      body: JSON.stringify(updatedData),
+    });
+  };
+
+
+  const closeTaskDetails = () => {
+    setIsEditing(false);
+    setSelectedTask(null);
+  };
 
 
   return (
     <TaskContext.Provider value={{
-      tasks, fetchTasks,
-      page, setPage,
-      selectedTask, setSelectedTask,
-      showDiscarded, setShowDiscarded
+      // states
+      tasks,
+      page,
+      selectedTask,
+      showDiscarded,
+      isEditing,
+
+      // functions
+      fetchTasks,
+      setPage,
+      setSelectedTask,
+      toggleDiscardedView,
+      setIsEditing,
+      updateTask,
+      closeTaskDetails
     }}>
       {children}
     </TaskContext.Provider>
